test/utils/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.6",
 "python-dateutil>=2.9.0.post0",
 "typer>=0.15.1",
]
